requirements.txt
PyMuPDF==1.26.3
numpy==2.3.2
google-generativeai==0.8.5
streamlit==1.48.1
scikit-learn==1.7.1